made the line five of unique_number_displayer.py file as float to accomodate decimal numbers in the program

+++ unique_number_displayer.py
@@ -2,7 +2,7 @@ from collections import Counter
 ten_numbers = []
 print("Enter ten numbers:")
 for i in range(10):
-    given_numbers = int(input(f"Number {i+1}: "))
+    given_numbers = float(input(f"Number {i+1}: "))
     ten_numbers.append(given_numbers)
 
 count = Counter(ten_numbers)
